feat(ui): Avoid circular dependency in routes files
Pull makeLazyloadComponent into its own file
