+++ static/gsApp/hooks/settingsRoutes.tsx
@@ -1,13 +1,10 @@
 import {Fragment} from 'react';
 
 import {IndexRedirect, IndexRoute, Redirect, Route} from 'sentry/components/route';
-import {makeLazyloadComponent} from 'sentry/routes';
+import {makeLazyloadComponent as make} from 'sentry/makeLazyloadComponent';
 import errorHandler from 'sentry/utils/errorHandler';
 
 import SubscriptionContext from 'getsentry/components/subscriptionContext';
-
-// Shorthand to avoid extra line wrapping
-const make = makeLazyloadComponent;
 
 const settingsRoutes = () =>
   (
